ask/requirements.txt
flask>=2.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
